+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.3.1
 
 - Removed `onlyIntegerCoords`
+- Removed `arrowClassName` from `LineContextProvider` props: Instead of setting `className` for every arrow, a better way would be to set a CSS selector for each path in the container (Example: `.svg-container path`)